% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_sheets_to_csv.R
\name{xl_sheets_to_csv}
\alias{xl_sheets_to_csv}
\title{Combine spreadsheets from excel workbooks into .csv files (one per workbook).}
\usage{
xl_sheets_to_csv(input_dir, output_dir = "./")
}
\arguments{
\item{input_dir}{String giving the directory containing the excel workbooks
to read from.}

\item{output_dir}{String giving the directory where to write .csv files to.}
}
\value{
Writes one .csv file for each workbook.
}
\description{
This is a rigid function with a very specific goal: To process
data from a specific sampling software. Specifically, this is what this
function does:
* Reads each spreadsheet from each workbook and map it to a dataframe.
* Lowercases and links the names of each dataframe.
* Keeps only these dataframes: (1) original_stems; (2) new_secondary_stems;
  (3) recruits; and (4) start_page.
* Keeps only non empty spreadsheets and warns what spreadsheets where dropped.
* Lowercases and links the names of each dataframe-variable.
* Drops fake stems.
* Joins all four dataframes into a single one.
* Output a single .csv file which name is prefixed with the name of the
  workbook.
}
\examples{
# Using sheets stored in the system.
path_to_example <- system.file("extdata", "example.xlsx", package = "fgeo.tool")
path_to_extdata <- fs::path(sub(basename(path_to_example), "", path_to_example))

# This should be the path to the directory containing the excel files
sheets_directory <- path_to_extdata
# This should be the path to the directory to output the .csv files
output_directory <- tempdir()
# Do the work
xl_sheets_to_csv(sheets_directory, output_directory)

# Confirm
fs::dir_ls(output_directory, regexp = "example.csv")
}
